manage notifications in browser.py with QSystrayicon
